+++ lib/src/messages/session.dart
@@ -4,7 +4,16 @@ part of '../../messages.dart';
     nullable: false, disallowUnrecognizedKeys: false, explicitToJson: true)
 class Session {
   final String id;
-  Session({@required this.id});
+
+  @JsonKey(nullable: true)
+  final String customer;
+
+  @JsonKey(name: 'payment_intent')
+  final String paymentIntent;
+  Session(
+      {@required this.id,
+      @required this.customer,
+      @required this.paymentIntent});
 
   factory Session.fromJson(Map<String, dynamic> json) =>
       _$SessionFromJson(json);