+++ lib/src/resources/payment_intent.dart
@@ -1,8 +1,11 @@
 import 'dart:async';
 
+import 'package:logging/logging.dart';
 import 'package:stripe/messages.dart';
 
 import '../client.dart';
+
+final log = Logger('Stripe PaymentIntentResource');
 
 class PaymentIntentResource {
   final Client _client;
@@ -11,5 +14,15 @@ class PaymentIntentResource {
   Future<PaymentIntent> retrieve(String paymentIntentId) async {
     final map = await _client.get(['payment_intents', paymentIntentId]);
     return PaymentIntent.fromJson(map);
+  }
+
+  Future<bool> cancel(String paymentIntentId) async {
+    try {
+      await _client.post(['payment_intents', paymentIntentId, 'cancel']);
+    } catch (e) {
+      log.warning(e);
+      return false;
+    }
+    return true;
   }
 }