+++ lib/src/resources/session.dart
@@ -35,4 +35,9 @@ class SessionResource {
     final response = await _client.post(['checkout', 'sessions'], data: data);
     return Session.fromJson(response);
   }
+
+  Future<Session> retrieve(String id) async {
+    final response = await _client.get(['checkout', 'sessions', id]);
+    return Session.fromJson(response);
+  }
 }